requirements.txt
sleekxmpp==1.3.1
redis==2.10.1
pyzmq==14.3.1
msgpack-python==0.4.2
